:art: - Moved /executions endpoint from Server.ts to api/exeutions.api.ts
